extract course table stuff to web/modules/course-table
