+++ modules/web/modules/course-table/empty-course-slot.js
