+++ modules/web/modules/course-table/empty-course-slot.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 .empty-course {
 	color: $gray-500;
