+++ modules/web/modules/course-table/index.js
@@ -1,0 +1,1 @@
+export { default } from './course-table-container'
