+++ modules/web/modules/course-table/semester.js
@@ -3,16 +3,16 @@ import { findDOMNode } from 'react-dom'
 import cx from 'classnames'
 import Link from 'react-router/lib/Link'
 import plur from 'plur'
-import { semesterName } from '../../../../../../school-st-olaf-college/course-info'
-import { countCredits } from '../../../../../../examine-student/count-credits'
-import { IDENT_COURSE } from '../../../../../../object-student/item-types'
+import { semesterName } from '../../../school-st-olaf-college/course-info'
+import { countCredits } from '../../../examine-student/count-credits'
+import { IDENT_COURSE } from '../../../object-student/item-types'
 import { DropTarget } from 'react-dnd'
 import includes from 'lodash/includes'
 
-import Button from '../../../../../components/button'
-import Icon from '../../../../../components/icon'
-import List from '../../../../../components/list'
-import { close, search } from '../../../../../icons/ionicons'
+import Button from '../../components/button'
+import Icon from '../../components/icon'
+import List from '../../components/list'
+import { close, search } from '../../icons/ionicons'
 
 import debug from 'debug'
 const log = debug('web:react')
