+++ modules/web/modules/course-table/semester.scss
@@ -1,5 +1,5 @@
-@import '../../../../../styles/variables.scss';
-@import '../../../../../styles/mixins.scss';
+@import '../../styles/variables.scss';
+@import '../../styles/mixins.scss';
 @import './variables.scss';
