+++ modules/web/modules/course-table/variables.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/variables.scss';
+@import '../../styles/variables.scss';
 
 $semester-spacing: 4px;
